+++ esp_cam_sensor/sensors/bf3a03/bf3a03.c
@@ -513,6 +513,7 @@ static esp_err_t bf3a03_power_on(esp_cam_sensor_device_t *dev)
 
     if (dev->xclk_pin >= 0) {
         BF3A03_ENABLE_OUT_XCLK(dev->xclk_pin, dev->xclk_freq_hz);
+        delay_ms(6);
     }
 
     if (dev->pwdn_pin >= 0) {