+++ esp_cam_sensor/sensors/gc2145/gc2145.c
@@ -833,6 +833,7 @@ static esp_err_t gc2145_power_on(esp_cam_sensor_device_t *dev)
 
     if (dev->xclk_pin >= 0) {
         GC2145_ENABLE_OUT_XCLK(dev->xclk_pin, dev->xclk_freq_hz);
+        delay_ms(5);
     }
 
     if (dev->pwdn_pin >= 0) {