+++ esp_cam_sensor/sensors/sc030iot/sc030iot.c
@@ -442,6 +442,7 @@ static esp_err_t sc030iot_power_on(esp_cam_sensor_device_t *dev)
 
     if (dev->xclk_pin >= 0) {
         SC030IOT_ENABLE_OUT_XCLK(dev->xclk_pin, dev->xclk_freq_hz);
+        delay_ms(6);
     }
 
     if (dev->pwdn_pin >= 0) {