+++ esp_video/examples/simple_video_server/main/simple_video_server_example.c
@@ -782,6 +782,10 @@ void app_main(void)
         ESP_ERROR_CHECK(nvs_flash_init());
     }
 
+    /*For camera devices that require the host to provide XCLK, the video_init() must be called immediately after the device is restarted,
+    otherwise the camera device may not be able to start due to the lack of the main clock.*/
+    ESP_ERROR_CHECK(example_video_init());
+
     ESP_ERROR_CHECK(esp_netif_init());
     ESP_ERROR_CHECK(esp_event_loop_create_default());
 
@@ -794,8 +798,6 @@ void app_main(void)
      * examples/protocols/README.md for more information about this function.
      */
     ESP_ERROR_CHECK(example_connect());
-
-    ESP_ERROR_CHECK(example_video_init());
 
     web_cam_video_config_t config[] = {
 #if EXAMPLE_ENABLE_MIPI_CSI_CAM_SENSOR