+++ esp_cam_sensor/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added line exposure time info in camera drivers.
 - Added OS02N10 MIPI camera sensor driver.
 - Added exposure value to "esp_cam_sensor_gh_exp_gain_t" to let the application set the exposure value instead of the exposure time
+- Added detect delay in DVP camera sensors driver.
 
 ## 1.3.0
 - Changed YUV422 sequence for SC030IOT, SC101IOT, GC0308, and BF3A03 sensors to depend on CONFIG_CAMERA_SENSOR_SWAP_PIXEL_BYTE_ORDER
@@ -12,7 +13,7 @@
 - Fixed OV2640 JPEG compression quality calculation issue.
 - Fixed incorrect format selection in OV2640 Kconfig.
 - Removed the GROUP Hold function in the SC2336 driver.
-- Modified GPIO pin type from int8_t to gpio_num_t.
+- Modified GPIO pin type from int8_t to gpio_num_t
 
 ## 1.2.1
 - Deprecated ledc intr type config