+++ esp_video/CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix sdkconfig.default.esp32s3 for better image bandwidth
 - Modify GPIO pin type from int8_t to gpio_num_t
 - Compatible with ESP-IDF v5.4.x(x ≥ 3), v5.5.1, and the later versions of the ESP-IDF DVP CAM driver
+- Initialize the video device first in the web example to avoid sensor detection failure.
 
 ## 1.1.0